requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Pillow==10.4.0
rembg==2.0.56
onnxruntime
